changed active learning sampler to use a sampler
