--- src/parsers/DATABASEparser.py
@@ -1,10 +1,0 @@
-from .base import Parser
-
-
-class DATABASEparser(Parser):
-    """ This is a dummy class as for the moment the sampler handles everything"""
-    def __init__(self, *args, **kwargs): 
-        pass
-
-    def write_input_file(self):
-        pass
