+++ src/runners/__init__.py
@@ -2,4 +2,4 @@ from .TGLFrunner import TGLFrunner
 from .SIMPLErunner import SIMPLErunner
 from .HELENArunner import HELENArunner
 from .MISHKArunner import MISHKArunner
-from .DATABASErunner import DATABASErunner
+from .STATICPOOLrunner import STATICPOOLrunner
